fix integer truncation bug

+++ service/src/mon.h
@@ -17,9 +17,9 @@ typedef struct app_struct {
     int pid;
     int tx;
     int rx;
-    uint8_t ctx;
-    uint8_t role;
-    uint8_t critical;
+    uint32_t ctx;
+    uint32_t role;
+    uint32_t critical;
     enum app_state state;
     char name[0x10];
     struct app_request cur_req;
